requirements.txt
flask>=2.0.1
flask-mysqldb>=1.0.1
PyMuPDF>=1.22.5
python-docx>=0.8.11
bcrypt>=4.0.1
torch>=2.0.1
transformers>=4.30.2
mysql-connector-python>=8.0.26 